requirements.txt
fastapi==0.111.0
pydantic>=2.7.0,<3.0.0
pytest~=8.2.2
SQLAlchemy~=2.0.31
#app~=0.0.1
pandas~=2.2.2
python-dotenv~=1.0.1
pydantic_core~=2.20.1
requests~=2.32.3
psycopg2~=2.9.9
psycopg2-binary~=2.9.9